requirements.txt
colorama==0.4.6
bcrypt==4.0.1
cryptography==43.0.1
python-dotenv